+++ src/components/BookItem.js
@@ -2,11 +2,15 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import { deleteBook } from '../redux/books/booksSlice';
+import DeleteBook from './DeleteBook';
 
 const BookItem = ({
   id, title, author,
 }) => {
   const dispatch = useDispatch();
+  const handleDelete = () => {
+    dispatch(deleteBook(id));
+  };
 
   return (
     <>
@@ -15,15 +19,7 @@ const BookItem = ({
           <h3>{title}</h3>
           <h4>{author}</h4>
         </div>
-        <button
-          type="button"
-          className="delete"
-          onClick={() => {
-            dispatch(deleteBook(id));
-          }}
-        >
-          Delete
-        </button>
+        <DeleteBook id={id} onDelete={handleDelete} />
       </div>
     </>
   );
